requirements.txt
arrow==0.17.0
asgiref==3.3.0
binaryornot==0.4.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
Django==3.1.3
django-phone-field==1.8.1
djangorestframework==3.12.2
idna==2.10
Jinja2==2.11.2
jinja2-time==0.2.0
MarkupSafe==1.1.1
poyo==0.5.0
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.4
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.25.11

factory-boy==2.12.0

pylint==2.4.2
flake8==3.7.9
flake8-black==0.1.1
black==19.10b0
django-prettyjson==0.4.1
python-tds==1.10.0
